opt: various arm speedups, a few fixes, and tiny avx2/avx512 speedups (via mainline)

+++ Source/Lib/Codec/mcomp.c
@@ -208,7 +208,7 @@ static INLINE unsigned int svt_check_better_fast(MacroBlockD *xd, const struct A
         if (var_params->bias_fp && (*best_mv).col % 8 == 0 && (*best_mv).row % 8 == 0)
             weight = BIAS_FP_WEIGHT;
 
-        if (((cost * weight) / 100) < *besterr) {
+        if ((((uint64_t)cost * weight) / 100) < *besterr) {
             *besterr    = cost;
             *best_mv    = *this_mv;
             *distortion = thismse;
@@ -239,7 +239,7 @@ static AOM_FORCE_INLINE unsigned int svt_check_better(MacroBlockD *xd, const str
         if (var_params->bias_fp && (*best_mv).col % 8 == 0 && (*best_mv).row % 8 == 0)
             weight = BIAS_FP_WEIGHT;
 
-        if (((cost * weight) / 100) < *besterr) {
+        if ((((uint64_t)cost * weight) / 100) < *besterr) {
             *besterr    = cost;
             *best_mv    = *this_mv;
             *distortion = thismse;
